# api/openapi.yaml
openapi: 3.0.3
info:
  title: NovaNAS Storage API
  version: 0.1.0
  description: ZFS-based storage control plane for a single-node NAS appliance.
servers:
  - url: /api/v1
paths:
...
  /scheduler/snapshot-schedules:
    get:
      operationId: listSnapshotSchedules
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SnapshotSchedule' }
    post:
      operationId: createSnapshotSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnapshotSchedule' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnapshotSchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /scheduler/snapshot-schedules/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getSnapshotSchedule
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnapshotSchedule' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateSnapshotSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnapshotSchedule' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SnapshotSchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteSnapshotSchedule
      responses:
        '204': { description: Deleted }
  /scheduler/replication-targets:
    get:
      operationId: listReplicationTargets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReplicationTarget' }
    post:
      operationId: createReplicationTarget
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplicationTarget' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplicationTarget' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /scheduler/replication-targets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getReplicationTarget
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplicationTarget' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteReplicationTarget
      responses:
        '204': { description: Deleted }
  /scheduler/replication-schedules:
    get:
      operationId: listReplicationSchedules
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReplicationSchedule' }
    post:
      operationId: createReplicationSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplicationSchedule' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplicationSchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /scheduler/replication-schedules/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getReplicationSchedule
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplicationSchedule' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateReplicationSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplicationSchedule' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplicationSchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      operationId: deleteReplicationSchedule
      responses:
        '204': { description: Deleted }
...
components:
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string }
        message: { type: string }
...
    SnapshotSchedule:
      type: object
      properties:
        id:               { type: string, format: uuid, readOnly: true }
        dataset:          { type: string }
        name:             { type: string }
        cron:             { type: string, description: 'Standard 5-field cron expression' }
        recursive:        { type: boolean }
        snapshotPrefix:   { type: string }
        retentionHourly:  { type: integer }
        retentionDaily:   { type: integer }
        retentionWeekly:  { type: integer }
        retentionMonthly: { type: integer }
        retentionYearly:  { type: integer }
        enabled:          { type: boolean }
    ReplicationTarget:
      type: object
      properties:
        id:            { type: string, format: uuid, readOnly: true }
        name:          { type: string }
        host:          { type: string }
        port:          { type: integer, default: 22 }
        sshUser:       { type: string }
        sshKeyPath:    { type: string }
        datasetPrefix: { type: string }
    ReplicationSchedule:
      type: object
      properties:
        id:              { type: string, format: uuid, readOnly: true }
        srcDataset:      { type: string }
        targetId:        { type: string, format: uuid }
        cron:            { type: string }
        snapshotPrefix:  { type: string }
        retentionRemote: { type: integer }
        enabled:         { type: boolean }
...