# internal/store/gen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package storedb

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type AuditLog struct {
	ID        int64              `json:"id"`
	Ts        pgtype.Timestamptz `json:"ts"`
	Actor     *string            `json:"actor"`
	Action    string             `json:"action"`
	Target    string             `json:"target"`
	RequestID string             `json:"request_id"`
	Payload   []byte             `json:"payload"`
	Result    string             `json:"result"`
}

type Job struct {
	ID         pgtype.UUID        `json:"id"`
	Kind       string             `json:"kind"`
	Target     string             `json:"target"`
	State      string             `json:"state"`
	Command    string             `json:"command"`
	Stdout     string             `json:"stdout"`
	Stderr     string             `json:"stderr"`
	ExitCode   *int32             `json:"exit_code"`
	Error      *string            `json:"error"`
	RequestID  string             `json:"request_id"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	StartedAt  pgtype.Timestamptz `json:"started_at"`
	FinishedAt pgtype.Timestamptz `json:"finished_at"`
}

type ReplicationSchedule struct {
	ID               pgtype.UUID        `json:"id"`
	SrcDataset       string             `json:"src_dataset"`
	TargetID         pgtype.UUID        `json:"target_id"`
	Cron             string             `json:"cron"`
	SnapshotPrefix   string             `json:"snapshot_prefix"`
	RetentionRemote  int32              `json:"retention_remote"`
	Enabled          bool               `json:"enabled"`
	LastFiredAt      pgtype.Timestamptz `json:"last_fired_at"`
	LastSyncSnapshot *string            `json:"last_sync_snapshot"`
	LastError        *string            `json:"last_error"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
}

type ReplicationTarget struct {
	ID            pgtype.UUID        `json:"id"`
	Name          string             `json:"name"`
	Host          string             `json:"host"`
	Port          int32              `json:"port"`
	SshUser       string             `json:"ssh_user"`
	SshKeyPath    string             `json:"ssh_key_path"`
	DatasetPrefix string             `json:"dataset_prefix"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
}

type ResourceMetadatum struct {
	ID          int64   `json:"id"`
	Kind        string  `json:"kind"`
	ZfsName     string  `json:"zfs_name"`
	DisplayName *string `json:"display_name"`
	Description *string `json:"description"`
	Tags        []byte  `json:"tags"`
}

type SnapshotSchedule struct {
	ID               pgtype.UUID        `json:"id"`
	Dataset          string             `json:"dataset"`
	Name             string             `json:"name"`
	Cron             string             `json:"cron"`
	Recursive        bool               `json:"recursive"`
	SnapshotPrefix   string             `json:"snapshot_prefix"`
	RetentionHourly  int32              `json:"retention_hourly"`
	RetentionDaily   int32              `json:"retention_daily"`
	RetentionWeekly  int32              `json:"retention_weekly"`
	RetentionMonthly int32              `json:"retention_monthly"`
	RetentionYearly  int32              `json:"retention_yearly"`
	Enabled          bool               `json:"enabled"`
	LastFiredAt      pgtype.Timestamptz `json:"last_fired_at"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
}